requirements.txt
rasterio
geopandas
scikit-learn
open-meteo
requests
requests-cache
retry-requests
python-dotenv